Rename remaining resnet.
